install typesync
