+++ package.json
@@ -7,17 +7,20 @@
     "build": "run-p type-check build-only",
     "preview": "vite preview",
     "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false"
+    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false",
+    "preinstall": "typesync || :"
   },
   "dependencies": {
     "idb-keyval": "^6.2.1",
     "vue": "^3.3.2",
-    "vue-router": "^4.2.0",
-    "vuex": "^4.0.2"
+    "vue-router": "^4.2.0"
   },
   "devDependencies": {
     "@tsconfig/node18": "^2.0.1",
+    "@types/bcrypt": "^10.0.28-alpha",
     "@types/node": "^18.16.8",
+    "@types/node-sass": "^4.11.7",
+    "@types/sass-loader": "^8.0.8",
     "@vitejs/plugin-vue": "^4.2.3",
     "@vue/tsconfig": "^0.4.0",
     "bcrypt": "^5.1.0",
@@ -26,7 +29,9 @@
     "sass": "^1.63.4",
     "sass-loader": "^13.3.2",
     "typescript": "~5.0.4",
+    "typesync": "^0.11.1",
     "vite": "^4.3.5",
-    "vue-tsc": "^1.6.4"
+    "vue-tsc": "^1.6.4",
+    "vuex": "^4.1.0"
   }
 }
